requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
mysql-connector==2.2.9
mysql-connector-python==9.0.0
python-dotenv==1.0.1
requests==2.32.3
tenacity==8.5.0
urllib3==2.2.2
yarl==1.9.4
